requirements.txt
Django==3.0.4
pycryptodome==3.9.7
